requirements.txt
streamlit>=1.24.0
tensorflow>=2.13.0
pillow
numpy